requirements.txt
httpx==0.25.2
openai==1.14.3
tiktoken==0.7.0
pydantic==2.7.0
python-dotenv==1.0.1
tqdm==4.66.2
backoff==2.2.1
rich==13.7.1
tokenizers==0.15.2
torch==2.3.0
llama_cpp_python==0.2.65